+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Changelog](https://keepachangelog.com/en/1.1.0).
     subcommand `cli-version`
   * Added CLI endpoint `init-tokens-mint` to mint all init tokens by
     spending the genesis UTxO.
+  * Added a `MinotaurStakeValidator` script.
 
 ## Changed
   * `init` now prints more human friendly message, when selected genesis utxo